code/requirements.txt
pandas
numpy
streamlit
plotly
matplotlib
seaborn
pydeck
openai
google-generativeai
pyspark
python-dotenv
groq
